requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-dotenv>=0.19.0
mlflow>=2.0.0
wandb>=0.15.0
prefect>=2.0.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
requests>=2.28.0
psutil>=5.9.0
prometheus-client>=0.15.0
structlog>=22.0.0
click>=8.0.0
pyyaml>=6.0 